+++ ncplib/errors.py
@@ -25,6 +25,9 @@ API reference
 .. autoexception:: DecodeError
     :members:
 
+.. autoclass:: Application
+    :members:
+
 .. autoexception:: CommandWarning
     :members:
 
@@ -38,9 +41,10 @@ __all__ = (
     "ConnectionError",
     "ConnectionClosed",
     "CommandError",
-    "CommandWarning",
-    "NCPWarning",
     "DecodeError",
+    "BadRequest",
+    "NCPWarning",
+    "CommandWarning",
     "DecodeWarning",
 )
 
@@ -110,6 +114,18 @@ class DecodeError(NCPError):
     """
     Raised when a non-recoverable error was encountered in a :term:`NCP packet`.
     """
+
+
+class BadRequest(NCPError):
+
+    """
+    An error that can be thrown in a field handler to signal a problem in handling the request.
+    """
+
+    def __init__(self, detail, code=400):
+        super().__init__("{detail!r} (code {code})".format(detail=detail, code=code))
+        self.detail = detail
+        self.code = code
 
 
 # Warnings.